Move methods to deprecated module
